LLS: Load title and summary from ext. package if avail.
Query for packages that support the ACTION_OPEN_LIVE_LOCK_SCREEN_SETTINGS
and load the title and summary for the preference from that package.
If no other packages exist or we are unable to load the title AND summary,
then we use the defaults and carry on as usual.

TICKET: CYNGNOS-2696

Change-Id: Ibbd5347aa35994877fa067b29cacbe9a9ed12dad